Moved files + fixed typo
