+++ packages/yourloops/components/patient-list/current-patient-list/current-patient-list.hook.tsx
@@ -34,24 +34,24 @@ import {
   type GridValueFormatterParams
 } from '@mui/x-data-grid'
 import { useTranslation } from 'react-i18next'
-import { usePatientListStyles } from './patient-list.styles'
-import { PatientListColumns } from './models/enums/patient-list.enum'
-import { usePatientContext } from '../../lib/patient/patient.provider'
-import { getMedicalValues } from '../patient/utils'
-import { ActionsCell, FlagIconCell, MessageCell, MonitoringAlertsCell } from './custom-cells'
-import { useAuth } from '../../lib/auth'
-import { type Patient } from '../../lib/patient/models/patient.model'
-import PatientUtils from '../../lib/patient/patient.util'
-import { type GridRowModel } from './models/grid-row.model'
-import { type UserToRemove } from '../dialogs/remove-direct-share-dialog'
+import { usePatientListStyles } from '../patient-list.styles'
+import { PatientListColumns } from '../models/enums/patient-list.enum'
+import { usePatientContext } from '../../../lib/patient/patient.provider'
+import { getMedicalValues } from '../../patient/utils'
+import { ActionsCell, FlagIconCell, MessageCell, MonitoringAlertsCell } from '../custom-cells'
+import { useAuth } from '../../../lib/auth'
+import { type Patient } from '../../../lib/patient/models/patient.model'
+import PatientUtils from '../../../lib/patient/patient.util'
+import { type GridRowModel } from '../models/grid-row.model'
+import { type UserToRemove } from '../../dialogs/remove-direct-share-dialog'
 import { formatBirthdate, getPatientFullName } from 'dumb'
 import { useNavigate } from 'react-router-dom'
 import Box from '@mui/material/Box'
-import { AppUserRoute } from '../../models/enums/routes.enum'
-import { sortByDateOfBirth, sortByFlag, sortByMonitoringAlertsCount, sortByUserName } from './sort-comparators.util'
-import { getUserName } from '../../lib/auth/user.util'
-import { useSelectedTeamContext } from '../../lib/selected-team/selected-team.provider'
-import { TeamType } from '../../lib/team/models/enums/team-type.enum'
+import { AppUserRoute } from '../../../models/enums/routes.enum'
+import { sortByDateOfBirth, sortByFlag, sortByMonitoringAlertsCount, sortByUserName } from '../sort-comparators.util'
+import { getUserName } from '../../../lib/auth/user.util'
+import { useSelectedTeamContext } from '../../../lib/selected-team/selected-team.provider'
+import { TeamType } from '../../../lib/team/models/enums/team-type.enum'
 
 interface CurrentPatientListHookProps {
   patients: Patient[]
